Simplify solution for Task v6 17.05 (Haskell)
Our first solution extracted a pair from the map and stored it in
oldRange, or set it to (index, index) if the key newLevel was not in
the map. However, the end of the range (i.e., the second number in the
pair) was never used because (fst oldRange, index) was the new range
that was written to the map.

This patch makes this a bit more elegant by storing only the start of
the range (i.e., the first number of the pair). This means that we only
need a default value for the range start if the map lookup returns
Nothing.